+++ R/randomCoefplot.R
@@ -9,6 +9,7 @@
 #' @param mar vector of length 4, which defines the margin sizes: \code{c(bottom, left, top, right)}. Defaults to \code{c(4,5,2,1)}.
 #' @param xlim.list list of vectors with length of two to define the intervals for x axis in each covariate plot. Defaults to NULL when the interval is defined by the range of point estimates and confidence intervals
 #' @param order logical, if \code{TRUE} (default), coefficients are sorted according to the point estimates
+#' @param ind.spp vector of species indices to construct the caterpillar plot for
 #' @param ...	additional graphical arguments.
 #'
 #' @author Jenni Niku <jenni.m.e.niku@@jyu.fi>, Francis K.C. Hui, Bert van der Veen, Sara Taskinen,
@@ -30,9 +31,9 @@
 #'@aliases randomCoefplot randomCoefplot.gllvm
 #'@export
 #'@export randomCoefplot.gllvm
-randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex.ylab = 0.5, mfrow = NULL, mar = c(4,6,2,1), xlim.list = NULL, order = FALSE, ...)
+randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex.ylab = 0.5, mfrow = NULL, mar = c(4,6,2,1), xlim.list = NULL, order = FALSE, ind.spp = NULL, ...)
 {
-  
+  if(is.null(ind.spp))ind.spp <- 1:ncol(object$y)
   if(is.null(object$lv.X.design) && !is.null(object$lv.X))object$lv.X.design <- object$lv.X # backward compatibility
   if (any(class(object) != "gllvm"))
     stop("Class of the object isn't 'gllvm'.")
@@ -45,16 +46,16 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex
   
   if((object$num.lv.c+object$num.RR)==0 && !is.null(object$params$Br) && !is.null(object$Xr)){
     if(is.null(which.Xcoef))which.Xcoef <- c(1:NROW(object$params$Br))
-    Xcoef <- as.matrix(t(object$params$Br)[,which.Xcoef,drop=F])
-    cnames <- colnames(object$Xr[,which.Xcoef,drop=FALSE])
+    Xcoef <- as.matrix(t(object$params$Br)[ind.spp,which.Xcoef,drop=F])
+    cnames <- colnames(object$Xr[ind.spp,which.Xcoef,drop=FALSE])
     k <- length(cnames)
-    if(is.null(colnames(object$y))) 
-      colnames(object$y) <- paste("Y",1:NCOL(object$y), sep = "")
-    m <- ncol(object$y)
+    if(is.null(colnames(object$y))) colnames(object$y) <- paste("Y",1:NCOL(object$y), sep = "")
+    labely <- colnames(object$y)[ind.spp]
+    m <- length(labely)
     Xc <- Xcoef
     
     sdXcoef <- t(predErr$Br)
-    sdXcoef <- sdXcoef[,which.Xcoef,drop=F]
+    sdXcoef <- sdXcoef[ind.spp,which.Xcoef,drop=F]
     if (is.null(mfrow) && k > 1)
       mfrow <- c(1, k)
     if (!is.null(mfrow))
@@ -83,7 +84,7 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex
       segments( x0 = lower, y0 = At.y, x1 = upper, y1 = At.y, col = col.seq )
       abline(v = 0, lty = 1)
       if (y.label)
-        axis( 2, at = At.y, labels = names(Xc), las = 1, cex.axis = cex.ylab)
+        axis( 2, at = At.y, labels = labely, las = 1, cex.axis = cex.ylab)
     }
   }
   
@@ -114,9 +115,9 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex
     k <- length(cnames)
     if(is.null(colnames(object$y))) 
       colnames(object$y) <- paste("Y",1:NCOL(object$y), sep = "")
-    labely <- colnames(object$y)
+    labely <- colnames(object$y)[ind.spp]
     m <- length(labely)
-    Xc <- Xcoef
+    sdXcoef <- sdXcoef[ind.spp,,drop=FALSE]
     
     if (is.null(mfrow) && k > 1)
       mfrow <- c(1, k)
@@ -125,7 +126,7 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex
     if (is.null(mfrow))
       par(mar = mar)
     for (i in 1:k) {
-      Xc <- Xcoef[, i]
+      Xc <- Xcoef[ind.spp, i,drop=FALSE]
       lower <- Xc - 1.96 * sdXcoef[, i]
       upper <- Xc + 1.96 * sdXcoef[, i]
       if(order){
@@ -146,7 +147,7 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, which.Xcoef = NULL, cex
       segments( x0 = lower, y0 = At.y, x1 = upper, y1 = At.y, col = col.seq )
       abline(v = 0, lty = 1)
       if (y.label)
-        axis( 2, at = At.y, labels = names(Xc), las = 1, cex.axis = cex.ylab)
+        axis( 2, at = At.y, labels = labely, las = 1, cex.axis = cex.ylab)
     }
   }
 }